+++ lib/govuk_personalisation.rb
@@ -3,6 +3,7 @@
 require "govuk_personalisation/version"
 require "govuk_personalisation/controller_concern"
 require "govuk_personalisation/flash"
+require "govuk_personalisation/redirect"
 require "govuk_personalisation/test_helpers/features"
 require "govuk_personalisation/test_helpers/requests"
 require "govuk_personalisation/urls"