+++ spec/test_app/app/controllers/redirects_controller.rb
@@ -1,0 +1,7 @@
+class RedirectsController < ApplicationController
+  include GovukPersonalisation::ControllerConcern
+
+  def show
+    redirect_with_analytics("https://example.com?keep=this")
+  end
+end